pyproject.toml
[tool.poetry]
name = "llm-operational-metrics"
version = "0.1.0"
description = ""
authors = ["Dennis Seah <dennis.seah@microsoft.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
lagom = "^2.6.0"
python-dotenv = "^1.0.1"
azure-identity = "^1.19.0"
azure-core = "^1.32.0"
openai = "^1.54.3"
azure-ai-inference = "^1.0.0b5"
aiohttp = "^3.10.10"
pydantic = "^2.9.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.2"
pre-commit = "^4.0.1"
pyright = "^1.1.388"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["typings"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "llm_operational_metrics/hosting.py"
]